添加了jwt认证及hutool工具 先解决用户的功能 之后考虑分模块

+++ src/main/java/com/freeLearn/wall/common/ResultCode.java
@@ -18,7 +18,7 @@ public enum ResultCode implements IErrorCode {
     }
 
     @Override
-    public long getCode() {
+    public int getCode() {
         return code;
     }
